check mistakes
