+++ js/data.js
@@ -26,57 +26,6 @@ const DESCRIPTIONS = [
   'Снимок на новый фотик',
 ];
 
-const EFFECTS = [
-  {
-    name: 'none',
-    style: 'none',
-    min: 0,
-    max: 100,
-    step: 1,
-    unit: '',
-  },
-  {
-    name: 'chrome',
-    style: 'grayscale',
-    min: 0,
-    max: 1,
-    step: 0.1,
-    unit: '',
-  },
-  {
-    name: 'sepia',
-    style: 'sepia',
-    min: 0,
-    max: 1,
-    step: 0.1,
-    unit: '',
-  },
-  {
-    name: 'marvin',
-    style: 'invert',
-    min: 0,
-    max: 100,
-    step: 1,
-    unit: '%',
-  },
-  {
-    name: 'phobos',
-    style: 'blur',
-    min: 0,
-    max: 3,
-    step: 0.1,
-    unit: 'px',
-  },
-  {
-    name: 'heat',
-    style: 'brightness',
-    min: 0,
-    max: 3,
-    step: 0.1,
-    unit: '',
-  },
-];
-
 const MESSAGES = [
   'Всё отлично!',
   'В целом всё неплохо. Но не всё.',
@@ -118,4 +67,4 @@ const getPhoto = () => {
 
 const getPhotos = () => Array.from({ length: ID_PHOTO_COUNT }, getPhoto);
 
-export { getPhotos, EFFECTS };
+export { getPhotos };
